+++ src/Item.php
@@ -4,45 +4,23 @@ namespace SebastiaanLuca\PipeOperator;
 
 class Item
 {
-    /**
-     * The current value being handled.
-     *
-     * @var mixed
-     */
-    protected $value;
+    protected mixed $value;
 
-    /**
-     * @param mixed $value The value you want to process.
-     */
-    public function __construct($value)
+    public function __construct(mixed $value)
     {
         $this->value = $value;
 
         if (! defined('PIPED_VALUE')) {
-            define('PIPED_VALUE', 'PIPED_VALUE-' . uniqid());
+            define('PIPED_VALUE', 'PIPED_VALUE-'.uniqid('', true));
         }
     }
 
-    /**
-     * @param string $name
-     * @param array $arguments
-     *
-     * @return mixed
-     */
-    public function __call($name, $arguments)
+    public function __call(string $name, array $arguments): mixed
     {
         return $this->pipe($name, ...$arguments);
     }
 
-    /**
-     * Perform an operation on the current value.
-     *
-     * @param callable|string|object $callback
-     * @param array ...$arguments
-     *
-     * @return \SebastiaanLuca\PipeOperator\Item|\SebastiaanLuca\PipeOperator\PipeProxy
-     */
-    public function pipe($callback, ...$arguments)
+    public function pipe(callable|object|string $callback, mixed ...$arguments): Item|PipeProxy
     {
         if (! is_callable($callback)) {
             return new PipeProxy($this, $callback);
@@ -55,14 +33,7 @@ class Item
         return $this;
     }
 
-    /**
-     * Add the given value to the list of arguments.
-     *
-     * @param  array $arguments
-     *
-     * @return array
-     */
-    public function addValueToArguments(array $arguments) : array
+    public function addValueToArguments(array $arguments): array
     {
         // If the caller hasn't explicitly specified where they want the value
         // to be added, we will add it as the first value. Otherwise we will
@@ -77,12 +48,7 @@ class Item
         }, $arguments);
     }
 
-    /**
-     * Get the current value.
-     *
-     * @return mixed
-     */
-    public function get()
+    public function get(): mixed
     {
         return $this->value;
     }