+++ tests/Unit/ObjectTest.php
@@ -10,7 +10,7 @@ class ObjectTest extends TestCase
     /**
      * @test
      */
-    public function it returns an item object when get has not been called yet() : void
+    public function it returns an item object when get has not been called yet(): void
     {
         $this->assertInstanceOf(
             Item::class,
@@ -21,7 +21,7 @@ class ObjectTest extends TestCase
     /**
      * @test
      */
-    public function it returns an item object when get has not been called yet using the method directly() : void
+    public function it returns an item object when get has not been called yet using the method directly(): void
     {
         $this->assertInstanceOf(
             Item::class,
@@ -32,7 +32,7 @@ class ObjectTest extends TestCase
     /**
      * @test
      */
-    public function it can transform a complex value in multiple steps() : void
+    public function it can transform a complex value in multiple steps(): void
     {
         $this->assertSame(
             'blog',
@@ -48,7 +48,7 @@ class ObjectTest extends TestCase
     /**
      * @test
      */
-    public function it can transform a complex value in multiple steps using the method directly() : void
+    public function it can transform a complex value in multiple steps using the method directly(): void
     {
         $this->assertSame(
             'blog',
@@ -63,7 +63,7 @@ class ObjectTest extends TestCase
     /**
      * @test
      */
-    public function it can transform a complex value in multiple steps using different method calls() : void
+    public function it can transform a complex value in multiple steps using different method calls(): void
     {
         $this->assertSame(
             'blog',