+++ tests/Unit/MethodsTest.php
@@ -4,13 +4,29 @@ namespace SebastiaanLuca\PipeOperator\Tests\Unit\Classes;
 
 use Closure;
 use PHPUnit\Framework\TestCase;
+use SebastiaanLuca\PipeOperator\Pipe;
 
 class MethodsTest extends TestCase
 {
     /**
      * @test
      */
-    public function it can transform a value using a callable string method() : void
+    public function it can transform using the static constructor(): void
+    {
+        $result = Pipe::from('https://blog.github.com')
+            ->parse_url()
+            ->end()
+            ->explode('.', PIPED_VALUE)
+            ->reset()
+            ->get();
+
+        $this->assertSame('blog', $result);
+    }
+
+    /**
+     * @test
+     */
+    public function it can transform a value using a callable string method(): void
     {
         $this->assertSame(
             'STRING',
@@ -23,7 +39,7 @@ class MethodsTest extends TestCase
     /**
      * @test
      */
-    public function it can transform a value using a callable string method using the method directly() : void
+    public function it can transform a value using a callable string method using the method directly(): void
     {
         $this->assertSame(
             'STRING',
@@ -36,13 +52,13 @@ class MethodsTest extends TestCase
     /**
      * @test
      */
-    public function it can transform a value using a closure() : void
+    public function it can transform a value using a closure(): void
     {
         $this->assertSame(
             'prefixed-string',
             take('string')
                 ->pipe(function (string $value) {
-                    return 'prefixed-' . $value;
+                    return 'prefixed-'.$value;
                 })
                 ->get()
         );
@@ -51,7 +67,20 @@ class MethodsTest extends TestCase
     /**
      * @test
      */
-    public function it can transform a value using a public class method() : void
+    public function it can transform a value using a short closure(): void
+    {
+        $this->assertSame(
+            'prefixed-string',
+            take('string')
+                ->pipe(fn (string $value): string => 'prefixed-'.$value)
+                ->get()
+        );
+    }
+
+    /**
+     * @test
+     */
+    public function it can transform a value using a public class method(): void
     {
         $this->assertSame(
             'UPPERCASE',
@@ -64,7 +93,7 @@ class MethodsTest extends TestCase
     /**
      * @test
      */
-    public function it can transform a value using a proxied public class method() : void
+    public function it can transform a value using a proxied public class method(): void
     {
         $this->assertSame(
             'UPPERCASE',
@@ -77,7 +106,7 @@ class MethodsTest extends TestCase
     /**
      * @test
      */
-    public function it can transform a value using a private class method() : void
+    public function it can transform a value using a private class method(): void
     {
         $this->assertSame(
             'lowercase',
@@ -90,7 +119,7 @@ class MethodsTest extends TestCase
     /**
      * @test
      */
-    public function it can transform a value using a proxied private class method() : void
+    public function it can transform a value using a proxied private class method(): void
     {
         $this->assertSame(
             'start-add-this',
@@ -104,7 +133,7 @@ class MethodsTest extends TestCase
     /**
      * @test
      */
-    public function it can transform a value while accepting pipe parameters() : void
+    public function it can transform a value while accepting pipe parameters(): void
     {
         $this->assertSame(
             ['KEY' => 'value'],
@@ -117,7 +146,7 @@ class MethodsTest extends TestCase
     /**
      * @test
      */
-    public function it can transform a value while accepting pipe parameters using the method directly() : void
+    public function it can transform a value while accepting pipe parameters using the method directly(): void
     {
         $this->assertSame(
             ['KEY' => 'value'],
@@ -127,32 +156,17 @@ class MethodsTest extends TestCase
         );
     }
 
-    /**
-     * @param string $value
-     *
-     * @return string
-     */
-    public function uppercase(string $value) : string
+    public function uppercase(string $value): string
     {
         return mb_strtoupper($value);
     }
 
-    /**
-     * @param string $value
-     *
-     * @return string
-     */
-    private function lowercase(string $value) : string
+    private function lowercase(string $value): string
     {
         return mb_strtolower($value);
     }
 
-    /**
-     * @param array ...$values
-     *
-     * @return string
-     */
-    private function join(...$values) : string
+    private function join(string ...$values): string
     {
         return implode('-', $values);
     }